+++ src/main/java/com/jobnote/auth/handler/CustomLogoutHandler.java
@@ -1,7 +1,6 @@
 package com.jobnote.auth.handler;
 
 import com.jobnote.domain.user.service.AuthTokenService;
-import com.jobnote.global.exception.JobNoteException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
@@ -12,9 +11,8 @@ import org.springframework.security.web.authentication.logout.LogoutHandler;
 import org.springframework.stereotype.Component;
 
 import static com.jobnote.global.common.Constants.*;
-import static com.jobnote.global.common.ResponseCode.INVALID_TOKEN;
-import static com.jobnote.global.util.CookieUtil.getCookie;
-import static com.jobnote.global.util.ResponseUtil.createResponseCookie;
+import static com.jobnote.global.util.CookieUtil.getTokenFromCookie;
+import static com.jobnote.global.util.CookieUtil.invalidateCookie;
 
 @RequiredArgsConstructor
 @Component
@@ -24,18 +22,10 @@ public class CustomLogoutHandler implements LogoutHandler {
 
     @Override
     public void logout(final HttpServletRequest request, final HttpServletResponse response, final Authentication authentication) {
-        final String refreshToken = getCookie(request, COOKIE_NAME_REFRESH_TOKEN)
-                .orElseThrow(() -> new JobNoteException(INVALID_TOKEN))
-                .getValue();
+        authTokenService.invalidate(getTokenFromCookie(request, COOKIE_NAME_REFRESH_TOKEN));
 
-        authTokenService.invalidate(refreshToken);
-
-        response.addHeader(HttpHeaders.SET_COOKIE,
-                createResponseCookie(COOKIE_NAME_ACCESS_TOKEN, null, "/", 0));
-
-        response.addHeader(HttpHeaders.SET_COOKIE,
-                createResponseCookie(COOKIE_NAME_REFRESH_TOKEN, null, "/", 0));
-
+        response.addHeader(HttpHeaders.SET_COOKIE, invalidateCookie(COOKIE_NAME_ACCESS_TOKEN));
+        response.addHeader(HttpHeaders.SET_COOKIE, invalidateCookie(COOKIE_NAME_REFRESH_TOKEN));
         SecurityContextHolder.clearContext();
     }
 }
