+++ src/main/java/com/jobnote/global/util/ResponseUtil.java
@@ -4,9 +4,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.jobnote.global.common.ApiResponse;
 import com.jobnote.global.common.ResponseCode;
 import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.boot.web.server.Cookie;
 import org.springframework.http.MediaType;
-import org.springframework.http.ResponseCookie;
 
 import java.io.IOException;
 
@@ -27,15 +25,5 @@ public class ResponseUtil {
         response.setContentType(MediaType.APPLICATION_JSON_VALUE);
         response.setCharacterEncoding(CHARACTER_ENCODING);
         response.getWriter().write(objectMapper.writeValueAsString(ApiResponse.ofSuccess(responseCode)));
-    }
-
-    public static String createResponseCookie(final String name, final String value, final String path, final int maxAge) {
-        return ResponseCookie.from(name, value)
-                .httpOnly(true)
-                .path(path)
-                .maxAge(maxAge)
-                // .secure(true) 운영 환경에서 설정
-                .sameSite(String.valueOf(Cookie.SameSite.STRICT))
-                .build().toString();
     }
 }
